Separated dynamic multiclass pipeline into training and evaluation scripts
